Fix for language files not being loaded
